# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de Reservas de Club Deportivo
  description: |
    API REST para gestionar canchas, socios y reservas del Club Deportivo Encuentro.

    El contrato sigue el enunciado del trabajo práctico. Todas las rutas se
    sirven bajo la URL base `/club_deportivo_api`.
  version: 1.0.0
servers:
  - url: http://localhost:5000/club_deportivo_api
    description: Servidor local de desarrollo

tags:
  - name: Deportes
    description: Deportes precargados
  - name: Canchas
    description: Gestión y disponibilidad de canchas
  - name: Socios
    description: Gestión de socios del club
  - name: Reservas
    description: Reservas de canchas

paths:
  /deportes:
    get:
      summary: Listar deportes precargados
      description: |
        Devuelve todos los deportes cargados por el script inicial. No usa
        paginación ni admite parámetros (cualquier parámetro responde 400).
      tags:
        - Deportes
      responses:
        '200':
          $ref: '#/components/responses/GetDeportesList200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /canchas:
    get:
      summary: Listar canchas
      description: |
        Devuelve el listado paginado de canchas. Se pueden aplicar filtros por
        deporte, nombre (búsqueda parcial e insensible a mayúsculas), techada y
        activa. Sin filtros se devuelven canchas activas e inactivas.
      tags:
        - Canchas
      parameters:
        - $ref: '#/components/parameters/IdDeporteFilter'
        - $ref: '#/components/parameters/NombreFilter'
        - $ref: '#/components/parameters/TechadaFilter'
        - $ref: '#/components/parameters/ActivaFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/GetCanchasList200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Crear una cancha
      tags:
        - Canchas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanchaCreate'
      responses:
        '201':
          $ref: '#/components/responses/CanchaCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /canchas/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Obtener una cancha por ID
      tags:
        - Canchas
      responses:
        '200':
          $ref: '#/components/responses/GetCancha200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Actualizar parcialmente una cancha
      tags:
        - Canchas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanchaUpdate'
      responses:
        '200':
          $ref: '#/components/responses/GetCancha200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Eliminar una cancha
      description: |
        Elimina una cancha solo si no tiene reservas asociadas. Si tiene
        reservas, responde 409 y podrá desactivarse mediante PATCH.
      tags:
        - Canchas
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /canchas/disponibles:
    get:
      summary: Consultar canchas disponibles en un intervalo
      description: |
        Devuelve las canchas activas que están libres durante todo el intervalo
        solicitado. No crea ni retiene la reserva.
      tags:
        - Canchas
      parameters:
        - $ref: '#/components/parameters/FechaRequired'
        - $ref: '#/components/parameters/HoraInicioRequired'
        - $ref: '#/components/parameters/HoraFinRequired'
        - $ref: '#/components/parameters/IdDeporteFilter'
        - $ref: '#/components/parameters/TechadaFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/GetCanchasList200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /socios:
    get:
      summary: Listar socios
      description: |
        Devuelve el listado paginado de socios con filtros opcionales por nombre
        (búsqueda parcial e insensible a mayúsculas) y activo.
      tags:
        - Socios
      parameters:
        - $ref: '#/components/parameters/NombreFilter'
        - $ref: '#/components/parameters/ActivoFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/GetSociosList200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Crear un socio
      tags:
        - Socios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocioCreate'
      responses:
        '201':
          $ref: '#/components/responses/SocioCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /socios/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Obtener un socio por ID
      tags:
        - Socios
      responses:
        '200':
          $ref: '#/components/responses/GetSocio200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Actualizar parcialmente un socio
      tags:
        - Socios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocioUpdate'
      responses:
        '200':
          $ref: '#/components/responses/GetSocio200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /reservas:
    get:
      summary: Listar reservas
      description: |
        Devuelve el listado paginado de reservas. Permite filtrar por cancha,
        socio, estado y rango de fechas (ambos extremos inclusive).
      tags:
        - Reservas
      parameters:
        - $ref: '#/components/parameters/IdCanchaFilter'
        - $ref: '#/components/parameters/IdSocioFilter'
        - $ref: '#/components/parameters/EstadoFilter'
        - $ref: '#/components/parameters/FechaDesdeFilter'
        - $ref: '#/components/parameters/FechaHastaFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/GetReservasList200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Crear una reserva
      description: |
        Crea una reserva en estado `confirmada`, validando que el socio y la
        cancha existan, estén activos y no tengan superposiciones. El servidor
        conserva la tarifa vigente y calcula el total.
      tags:
        - Reservas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaCreate'
      responses:
        '201':
          $ref: '#/components/responses/ReservaCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /reservas/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Obtener una reserva por ID
      tags:
        - Reservas
      responses:
        '200':
          $ref: '#/components/responses/GetReserva200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /reservas/{id}/estado:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      summary: Establecer el estado de una reserva
      description: |
        Cambia el estado de una reserva y devuelve la reserva con su estado actual.
        - `confirmada` → `cancelada`: solo si el horario de inicio todavía no llegó.
        - `confirmada` → `finalizada`: solo si se alcanzó o superó el horario de fin.
        - Una reserva `cancelada` o `finalizada` no puede cambiar de estado (409).
        - Repetir el estado actual devuelve 200 sin modificar la reserva.
        - Un estado desconocido responde 400.
      tags:
        - Reservas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstadoReserva'
      responses:
        '200':
          $ref: '#/components/responses/GetReserva200'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    Deporte:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único del deporte
          example: 1
        nombre:
          type: string
          description: Nombre del deporte
          example: "Fútbol"

    Cancha:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único de la cancha
          example: 1
        nombre:
          type: string
          description: Nombre o denominación de la cancha
          example: "Cancha 1 - Fútbol 5"
        id_deporte:
          type: integer
          description: Identificador del deporte asociado
          example: 1
        precio_hora:
          type: integer
          description: Tarifa vigente por hora en centavos
          example: 1000000
        techada:
          type: boolean
          description: Indica si la cancha es cubierta
          example: false
        activa:
          type: boolean
          description: Indica si admite nuevas reservas
          example: true

    CanchaCreate:
      type: object
      additionalProperties: false
      required:
        - nombre
        - id_deporte
        - precio_hora
      properties:
        nombre:
          type: string
          description: Nombre de la cancha
          example: "Cancha 1 - Fútbol 5"
        id_deporte:
          type: integer
          description: Identificador del deporte asociado
          example: 1
        precio_hora:
          type: integer
          description: Tarifa por hora en centavos. Debe ser mayor a cero.
          minimum: 1
          example: 1000000
        techada:
          type: boolean
          description: Indica si la cancha es cubierta
          default: false
          example: false
        activa:
          type: boolean
          description: Indica si la cancha admite nuevas reservas
          default: true
          example: true

    CanchaUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nombre:
          type: string
          description: Nombre de la cancha
          example: "Cancha 1 - Fútbol 5"
        precio_hora:
          type: integer
          description: Tarifa por hora en centavos. Debe ser mayor a cero.
          minimum: 1
          example: 1200000
        techada:
          type: boolean
          description: Indica si la cancha es cubierta
          example: true
        activa:
          type: boolean
          description: Indica si la cancha admite nuevas reservas
          example: true

    Socio:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único del socio
          example: 1
        nombre:
          type: string
          description: Nombre completo del socio
          example: "Juan Pérez"
        email:
          type: string
          format: email
          description: Correo electrónico único del socio
          example: "juan.perez@example.com"
        activo:
          type: boolean
          description: Indica si el socio puede realizar nuevas reservas
          example: true

    SocioCreate:
      type: object
      additionalProperties: false
      required:
        - nombre
        - email
      properties:
        nombre:
          type: string
          maxLength: 60
          description: Nombre completo del socio (se guardan sin espacios en los extremos)
          example: "Juan Pérez"
        email:
          type: string
          format: email
          maxLength: 80
          description: Correo electrónico del socio (se guarda en minúsculas y sin espacios en los extremos)
          example: "juan.perez@example.com"

    SocioUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nombre:
          type: string
          maxLength: 60
          description: Nombre completo del socio (se guardan sin espacios en los extremos)
          example: "Juan Pérez"
        email:
          type: string
          format: email
          maxLength: 80
          description: Correo electrónico del socio (se guarda en minúsculas y sin espacios en los extremos)
          example: "juan.perez@example.com"
        activo:
          type: boolean
          description: Indica si el socio puede realizar nuevas reservas
          example: true

    Reserva:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único de la reserva
          example: 1
        id_socio:
          type: integer
          description: Identificador del socio responsable
          example: 1
        id_cancha:
          type: integer
          description: Identificador de la cancha reservada
          example: 2
        fecha_hora_inicio:
          type: string
          format: date-time
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{6}-03:00$'
          description: Inicio del intervalo reservado en GMT-3 con 6 decimales de precisión
          example: "2026-10-15T18:00:00.000000-03:00"
        fecha_hora_fin:
          type: string
          format: date-time
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{6}-03:00$'
          description: Fin del intervalo reservado en GMT-3 con 6 decimales de precisión
          example: "2026-10-15T20:00:00.000000-03:00"
        estado:
          type: string
          description: Estado de la reserva
          enum:
            - confirmada
            - cancelada
            - finalizada
          example: "confirmada"
        precio_hora:
          type: integer
          description: Tarifa histórica aplicada en centavos
          example: 1000000
        precio_total:
          type: integer
          description: Importe total calculado en centavos
          example: 2000000

    ReservaCreate:
      type: object
      additionalProperties: false
      required:
        - id_socio
        - id_cancha
        - fecha_hora_inicio
        - fecha_hora_fin
      properties:
        id_socio:
          type: integer
          description: Identificador del socio responsable
          example: 1
        id_cancha:
          type: integer
          description: Identificador de la cancha reservada
          example: 2
        fecha_hora_inicio:
          type: string
          format: date-time
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{6}-03:00$'
          description: Inicio del intervalo reservado en GMT-3 con 6 decimales de precisión
          example: "2026-10-15T18:00:00.000000-03:00"
        fecha_hora_fin:
          type: string
          format: date-time
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{6}-03:00$'
          description: Fin del intervalo reservado en GMT-3 con 6 decimales de precisión
          example: "2026-10-15T20:00:00.000000-03:00"

    EstadoReserva:
      type: object
      additionalProperties: false
      required:
        - estado
      properties:
        estado:
          type: string
          description: Nuevo estado de la reserva
          enum:
            - confirmada
            - cancelada
            - finalizada
          example: "cancelada"

    Links:
      type: object
      description: Enlaces HATEOAS para navegar entre páginas
      properties:
        _first:
          $ref: '#/components/schemas/Link'
        _prev:
          $ref: '#/components/schemas/Link'
        _next:
          $ref: '#/components/schemas/Link'
        _last:
          $ref: '#/components/schemas/Link'

    Link:
      type: object
      properties:
        href:
          type: string
          description: URI para navegar a la página indicada
          example: "http://localhost:5000/canchas?_offset=0&_limit=10"

    Error:
      type: object
      properties:
        code:
          type: string
          description: Código alfanumérico del error
          example: "ERROR_VALIDACION"
        message:
          type: string
          description: Resumen breve del error
          example: "El cuerpo de la solicitud es inválido"
        level:
          type: string
          description: Nivel del error
          enum:
            - info
            - warning
            - error
          example: "error"
        description:
          type: string
          description: Descripción detallada del error
          example: "El campo 'precio_hora' debe ser un entero mayor a cero"

    Errores:
      type: object
      properties:
        errors:
          type: array
          description: Listado de errores
          items:
            $ref: '#/components/schemas/Error'

    DeportesListResponse:
      type: object
      properties:
        deportes:
          type: array
          items:
            $ref: '#/components/schemas/Deporte'

    CanchasListResponse:
      type: object
      properties:
        canchas:
          type: array
          items:
            $ref: '#/components/schemas/Cancha'
        _links:
          $ref: '#/components/schemas/Links'

    SociosListResponse:
      type: object
      properties:
        socios:
          type: array
          items:
            $ref: '#/components/schemas/Socio'
        _links:
          $ref: '#/components/schemas/Links'

    ReservasListResponse:
      type: object
      properties:
        reservas:
          type: array
          items:
            $ref: '#/components/schemas/Reserva'
        _links:
          $ref: '#/components/schemas/Links'

  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Identificador entero positivo del recurso
      schema:
        type: integer
      example: 1

    Limit:
      name: _limit
      in: query
      description: Cantidad máxima de registros por página
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10

    Offset:
      name: _offset
      in: query
      description: Desplazamiento desde el primer registro
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0

    NombreFilter:
      name: nombre
      in: query
      description: Búsqueda parcial e insensible a mayúsculas por nombre
      schema:
        type: string
      example: "Cancha 1"

    IdDeporteFilter:
      name: id_deporte
      in: query
      description: Filtrar por identificador de deporte
      schema:
        type: integer
      example: 1

    TechadaFilter:
      name: techada
      in: query
      description: Filtrar por indicador de cancha techada
      schema:
        type: boolean
      example: true

    ActivaFilter:
      name: activa
      in: query
      description: Filtrar por indicador de cancha activa
      schema:
        type: boolean
      example: true

    ActivoFilter:
      name: activo
      in: query
      description: Filtrar por indicador de socio activo
      schema:
        type: boolean
      example: true

    EstadoFilter:
      name: estado
      in: query
      description: Filtrar por estado de la reserva
      schema:
        type: string
        enum:
          - confirmada
          - cancelada
          - finalizada
      example: "confirmada"

    IdCanchaFilter:
      name: id_cancha
      in: query
      description: Filtrar por identificador de cancha
      schema:
        type: integer
      example: 2

    IdSocioFilter:
      name: id_socio
      in: query
      description: Filtrar por identificador de socio
      schema:
        type: integer
      example: 1

    FechaDesdeFilter:
      name: fecha_desde
      in: query
      description: Fecha inicial del rango inclusive
      schema:
        type: string
        format: date
      example: "2026-10-01"

    FechaHastaFilter:
      name: fecha_hasta
      in: query
      description: Fecha final del rango inclusive
      schema:
        type: string
        format: date
      example: "2026-10-31"

    FechaRequired:
      name: fecha
      in: query
      required: true
      description: Día para consultar disponibilidad
      schema:
        type: string
        format: date
      example: "2026-10-15"

    HoraInicioRequired:
      name: hora_inicio
      in: query
      required: true
      description: Hora de inicio del intervalo
      schema:
        type: string
        pattern: '^([01]\d|2[0-3]):00:00$'
      example: "18:00:00"

    HoraFinRequired:
      name: hora_fin
      in: query
      required: true
      description: Hora de fin del intervalo
      schema:
        type: string
        pattern: '^([01]\d|2[0-3]):00:00$'
      example: "20:00:00"

  responses:
    GetDeportesList200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DeportesListResponse'

    GetCanchasList200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CanchasListResponse'

    GetCancha200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Cancha'

    GetSociosList200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SociosListResponse'

    GetSocio200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Socio'

    GetReservasList200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReservasListResponse'

    GetReserva200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Reserva'

    CanchaCreated:
      description: Cancha creada. El cuerpo es vacío; la URL de la cancha viene en el header Location.
      headers:
        Location:
          description: URL de la cancha creada
          schema:
            type: string
            example: "http://localhost:5000/club_deportivo_api/canchas/7"

    SocioCreated:
      description: Socio creado. Se devuelve el socio con el id asignado y activo en true.
      headers:
        Location:
          description: URL del socio creado
          schema:
            type: string
            example: "http://localhost:5000/club_deportivo_api/socios/3"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Socio'

    ReservaCreated:
      description: Reserva creada en estado confirmada, con la tarifa vigente y el total calculado.
      headers:
        Location:
          description: URL de la reserva creada
          schema:
            type: string
            example: "http://localhost:5000/club_deportivo_api/reservas/5"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Reserva'

    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'

    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'

    Conflict:
      description: Conflicto de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'

    NoContent:
      description: Sin contenido

    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'